+++ basic/controllers/MarkrecordController.php
@@ -7,7 +7,8 @@ use app\models\MarkrecordSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-
+use Yii;
+use app\models\User;
 /**
  * MarkrecordController implements the CRUD actions for Markrecord model.
  */
@@ -30,7 +31,16 @@ class MarkrecordController extends Controller
             ]
         );
     }
-
+    public function checkadmin(){
+        if(Yii::$app->user->identity==null){
+            return false;
+        }
+        $user=User::findOne(['user_id'=>Yii::$app->user->identity->user_id]);
+        if($user->user_type==="admin"){
+            return true;
+        }
+        return false;
+    }
     /**
      * Lists all Markrecord models.
      *
@@ -38,6 +48,9 @@ class MarkrecordController extends Controller
      */
     public function actionIndex()
     {
+        if($this->checkadmin()===false){
+            return $this->redirect(['site/login']);
+        }
         $this->layout="backend";
         $searchModel = new MarkrecordSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
@@ -57,6 +70,9 @@ class MarkrecordController extends Controller
      */
     public function actionView($mark_id, $post_id)
     {
+        if($this->checkadmin()===false){
+            return $this->redirect(['site/login']);
+        }
         $this->layout="backend";
         return $this->render('view', [
             'model' => $this->findModel($mark_id, $post_id),
@@ -70,6 +86,9 @@ class MarkrecordController extends Controller
      */
     public function actionCreate()
     {
+        if($this->checkadmin()===false){
+            return $this->redirect(['site/login']);
+        }
         $this->layout="backend";
         $model = new Markrecord();
 
@@ -96,6 +115,9 @@ class MarkrecordController extends Controller
      */
     public function actionUpdate($mark_id, $post_id)
     {
+        if($this->checkadmin()===false){
+            return $this->redirect(['site/login']);
+        }
         $this->layout="backend";
         $model = $this->findModel($mark_id, $post_id);
 
@@ -118,6 +140,9 @@ class MarkrecordController extends Controller
      */
     public function actionDelete($mark_id, $post_id)
     {
+        if($this->checkadmin()===false){
+            return $this->redirect(['site/login']);
+        }
         $this->layout="backend";
         $this->findModel($mark_id, $post_id)->delete();
